requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
pip-autoremove==0.10.0
psycopg2-binary==2.9.10
setuptools==80.4.0
sqlparse==0.5.3
tzdata==2025.2
